requirements.txt
anthropic>=0.40.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langgraph>=0.2.0
langsmith>=0.1.0
chromadb>=0.5.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
sqlglot>=23.0.0
streamlit>=1.35.0
